Remove iex_try!
This macro is unsound. It applies even inside closures, allowing the
user to catch an IexPanic, which can then be rethrown in another
context, making iex believe an exception is in progress when it's not.

We would ideally make iex_try! sound by limiting its applicability, but
I'm not sure how to do that.